requirements.txt
Django==1.7.7
six==1.9.0
python-dateutil==2.4.2
Markdown==2.6.1
pyenchant==1.6.6
smartypants==1.8.6
uuid==1.30
pytz==2015.2
amqp==1.4.6
billiard==3.3.0.19
amqplib==1.0.2
anyjson==0.3.3
kombu==3.0.23
django-picklefield==0.3.1
pyparsing==2.0.3
celery==3.1.17
django-celery==3.1.16
poster==0.7.0
html5lib==0.999
BeautifulSoup==3.2.1
psycopg2==2.6
requests==2.6.0
coverage==3.7.1
pep8==1.5.7
pyflakes==0.8.1
mccabe==0.3
flake8==2.4.0
statsd==3.1
gunicorn==19.3.0
boto==2.34.0

template_utils==0.4p2
typogrify==2.0.7
django-statsd-mozilla==0.3.14
pycrypto==2.6.1
ecdsa==0.13
paramiko==1.15.2
Fabric==1.10.1
raven==5.2.0
django-markwhat==1.3
django-storages-redux==1.2.2
django-cacheds3storage==0.1.1
django-appconf==0.6
django-compressor==1.5
